Search Engine Optimization
SEO(Search Engine Optimization) is a marketing discipline focused on growing visibility in organic search engine results. In SIPL training institute we trained you in SEO encompasses both the technical creative elements required to improve rankings, drive traffic in search engines for your website.
SEO is short for Search Engine Optimization is nothing really mystical about it You might have heard a lot about SEO and how it works, but basically, it is is a measurable, repeatable process that is used to send signals search engines that your pages are worth showing in Google's index.